pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "fast-django-asgi"
version = "0.1.3"
description = "Django-like developer experience on ASGI powered by FastAPI, Tortoise ORM, Aerich, and FastAPI-Admin."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Aakar Sharma" }]
keywords = ["fastapi", "django", "tortoise-orm", "aerich", "admin", "asgi"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
  "Framework :: AsyncIO",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
]

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "tortoise-orm>=0.20",
  "aerich>=0.7.2",
  "fastapi-admin>=1.0.4",
  "typer>=0.12.3",
  "jinja2>=3.1",
  "pydantic>=2.5",
  "pydantic-settings>=2.2.1",
  "python-multipart>=0.0.9",
  "email-validator>=2.1.0",
  "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "anyio>=4.4",
]
dev = [
  "ruff>=0.5",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "types-PyYAML",
  "ipython",
]
docs = [
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
  "mkdocs-include-markdown-plugin>=6.0",
  "mkdocstrings[python]>=0.24",
]

[project.scripts]
fast-django = "fast_django.cli.main:app"

[tool.hatch.build]
packages = ["src/fast_django"]
include = [
  "src/fast_django/scaffolding/**",
  "src/fast_django/py.typed",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL", "TID", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["fast_django"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_any_generics = true
show_error_codes = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["passlib.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --cov=src/fast_django --cov-report=term-missing --cov-fail-under=90"
testpaths = ["tests"]
